More changes for CLS compliance.
Former-commit-id: 24bb7cd8527b2628b5a0634a4a70a376e2b9bdd6

+++ src/RiakClient/Comms/RiakPbcSocket.cs
@@ -133,7 +133,7 @@ namespace RiakClient.Comms
             {
                 var error = DeserializeInstance<RpbErrorResp>(size);
                 string errorMessage = error.errmsg.FromRiakString();
-                throw new RiakException(error.errcode, errorMessage, false);
+                throw new RiakException((int)error.errcode, errorMessage, false);
             }
 
             if (expectedCode != messageCode)
@@ -156,7 +156,7 @@ namespace RiakClient.Comms
             {
                 var error = DeserializeInstance<RpbErrorResp>(size);
                 string errorMessage = error.errmsg.FromRiakString();
-                throw new RiakException(error.errcode, errorMessage, false);
+                throw new RiakException((int)error.errcode, errorMessage, false);
             }
 
             if (!MessageCodeTypeMapBuilder.Contains(messageCode))
